+++ CHANGELOG.md
@@ -16,11 +16,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add details to mention filter log message.
 - Reduce number of database queries made during mention filtering.
-- Log `EmojiReact` and `Like` activities when `content` property is present.
 - Set timeout on monero-wallet-rpc requests.
 - Apply custom migrations when `mitractl` is used.
 - Copy instance RSA key from `instance_rsa_key` file to database.
 - Enable integrity proofs with `eddsa-jcs-2022` cryptosuite by default.
+- Fetch custom emojis used in `Like` activities.
 
 ## [2.11.0] - 2024-02-20